refactor(appointment): 약속잡기 결과 페이지 내부 컴포넌트들 추상화 레벨 조정
